added camera

+++ app/src/main/java/com/roxx/grigarage/data/local/model/BeerEntity.kt
@@ -12,7 +12,7 @@ data class BeerEntity(
     val alcoholPercentage: Double = 0.0,
     val volume: Float = 0f,
     val color: String = "",
-    val rating: Float = 0f,
+    val rating: Double = 0.0,
     val notes: String = "",
     val photoUri: String = "",
     val dateAdded: Long = 0,
